+++ packages/cosmwasm/src/index.ts
@@ -14,7 +14,6 @@ export {
   CosmWasmClient,
   GetSequenceResult,
   SearchByHeightQuery,
-  SearchByIdQuery,
   SearchBySentFromOrToQuery,
   SearchByTagsQuery,
   SearchTxQuery,