+++ packages/launchpad/src/cosmosclient.searchtx.spec.ts
@@ -26,7 +26,7 @@ interface TestTxSend {
   readonly tx: WrappedStdTx;
 }
 
-describe("CosmosClient.searchTx", () => {
+describe("CosmosClient.getTx and .searchTx", () => {
   let sendUnsuccessful: TestTxSend | undefined;
   let sendSuccessful: TestTxSend | undefined;
 
@@ -91,14 +91,13 @@ describe("CosmosClient.searchTx", () => {
     }
   });
 
-  describe("with SearchByIdQuery", () => {
-    it("can search successful tx by ID", async () => {
+  describe("getTx", () => {
+    it("can get successful tx by ID", async () => {
       pendingWithoutLaunchpad();
       assert(sendSuccessful, "value must be set in beforeAll()");
       const client = new CosmosClient(launchpad.endpoint);
-      const result = await client.searchTx({ id: sendSuccessful.hash });
-      expect(result.length).toEqual(1);
-      expect(result[0]).toEqual(
+      const result = await client.getTx(sendSuccessful.hash);
+      expect(result).toEqual(
         jasmine.objectContaining({
           height: sendSuccessful.height,
           hash: sendSuccessful.hash,
@@ -108,13 +107,12 @@ describe("CosmosClient.searchTx", () => {
       );
     });
 
-    it("can search unsuccessful tx by ID", async () => {
+    it("can get unsuccessful tx by ID", async () => {
       pendingWithoutLaunchpad();
       assert(sendUnsuccessful, "value must be set in beforeAll()");
       const client = new CosmosClient(launchpad.endpoint);
-      const result = await client.searchTx({ id: sendUnsuccessful.hash });
-      expect(result.length).toEqual(1);
-      expect(result[0]).toEqual(
+      const result = await client.getTx(sendUnsuccessful.hash);
+      expect(result).toEqual(
         jasmine.objectContaining({
           height: sendUnsuccessful.height,
           hash: sendUnsuccessful.hash,
@@ -124,39 +122,12 @@ describe("CosmosClient.searchTx", () => {
       );
     });
 
-    it("can search by ID (non existent)", async () => {
+    it("can get by ID (non existent)", async () => {
       pendingWithoutLaunchpad();
       const client = new CosmosClient(launchpad.endpoint);
       const nonExistentId = "0000000000000000000000000000000000000000000000000000000000000000";
-      const result = await client.searchTx({ id: nonExistentId });
-      expect(result.length).toEqual(0);
-    });
-
-    it("can search by ID and filter by minHeight", async () => {
-      pendingWithoutLaunchpad();
-      assert(sendSuccessful);
-      const client = new CosmosClient(launchpad.endpoint);
-      const query = { id: sendSuccessful.hash };
-
-      {
-        const result = await client.searchTx(query, { minHeight: 0 });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { minHeight: sendSuccessful.height - 1 });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { minHeight: sendSuccessful.height });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { minHeight: sendSuccessful.height + 1 });
-        expect(result.length).toEqual(0);
-      }
+      const result = await client.getTx(nonExistentId);
+      expect(result).toBeNull();
     });
   });