+++ packages/launchpad/src/index.ts
@@ -18,13 +18,11 @@ export {
   BroadcastTxResult,
   BroadcastTxSuccess,
   SearchByHeightQuery,
-  SearchByIdQuery,
   SearchBySentFromOrToQuery,
   SearchByTagsQuery,
   SearchTxQuery,
   SearchTxFilter,
   isSearchByHeightQuery,
-  isSearchByIdQuery,
   isSearchBySentFromOrToQuery,
   isSearchByTagsQuery,
 } from "./cosmosclient";